Fixed listTopConsumedResources to work w/o specifying page/pageSize

+++ server/src/com/cloud/capacity/dao/CapacityDao.java
@@ -40,7 +40,7 @@ public interface CapacityDao extends GenericDao<CapacityVO, Long> {
     Pair<List<Long>, Map<Long, Double>> orderPodsByAggregateCapacity(long zoneId, short capacityType, float cpuOverprovisioningFactor);
     List<SummedCapacity> findCapacityBy(Integer capacityType, Long zoneId,
             Long podId, Long clusterId, String resourceState);
-    List<SummedCapacity> listCapacitiesGroupedByLevelAndType(Integer capacityType, Long zoneId, Long podId, Long clusterId, int level, Integer limit);  
+    List<SummedCapacity> listCapacitiesGroupedByLevelAndType(Integer capacityType, Long zoneId, Long podId, Long clusterId, int level, Long limit);  
     void updateCapacityState(Long dcId, Long podId, Long clusterId,
             Long hostId, String capacityState);
 }
